+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.12.2] - 2024-01-05
+
+### Fixed
+
+* Stopping a scheduled task does not stop other tasks with the same execution time.
+
+## [1.12.1] - 2023-10-29
+
+### Added
+
+* UuidUtils.add method.
+
 ## [1.12.1] - 2023-10-29
 
 ### Added